"fix" for the scaling bug
The "fix" is to write the data to csv and read it from csv before scaling. I suspect this truncates decimals and prevents weird floating point precision errors in sklearn (possibly a bug in sklearn because the errors are different depending on how StandardScaler is applied). Currently considering whether this fix should be applied to within the scale_data function... It would decrease efficiency but reduce the amount of things the user has to consider.